[이지은] feat: 비밀번호 찾기 시스템 구현중 [FPS]

+++ jieunLee/backend/src/main/java/com/example/backend/controller/BoardController.java
@@ -62,7 +62,7 @@ public class BoardController {
 
     @DeleteMapping("/{boardNo}")
     public void boardRemove (
-            @PathVariable("boardNo") Integer boardNo) {
+            @PathVariable("boardNo") Long boardNo) {
         log.info("boardRemove()");
 
         service.remove(boardNo);
